requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai>=1.75.0
chromadb==0.6.3
python-telegram-bot==21.10
python-dotenv==1.0.1
